requirements.txt
appdirs==1.4.3
defusedxml==0.5.0
Django==1.11
django-allauth==0.32.0
django-constance==2.0.0
django-crispy-forms==1.6.1
django-csv-exports==1.0.3
django-dynamic-fixture==1.9.4
django-jsonfield==1.0.1
django-modeltranslation==0.12.1
django-picklefield==0.3.2
django-rosetta==0.7.13
django-summernote==0.8.6
microsofttranslator==0.8
oauthlib==2.0.2
olefile==0.44
packaging==16.8
Pillow==4.1.0
polib==1.0.8
pyparsing==2.2.0
python3-openid==3.1.0
pytz==2017.2
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
sorl-thumbnail==12.3
